[ 3403] Add script to update MD5-based scrambled passwords in catalog
This script re-scrambles passwords stored in the iRODS catalog which
use MD5 as part of the scrambling algorithm to instead use SHA256.

It includes a dry-run option to just count the rows which will be updated.

It also includes an option to specify the input and output hash schemes
so that passwords can be scrambled back and forth between schemes. The
default option values are an input scheme of MD5 and an output scheme
of SHA256.