[ XXXX] Add hashed password support
This commit adds hashed password support for authentication in iRODS.
No existing authentication scheme uses hashed passwords, but this just
adds the feature for later use.

A new database table called R_USER_CREDENTIALS has been added for
storing hashed passwords. A new database operation has been added for
checking the user's password against the hashed passwords in the new
table.

A server configuration has been added that allows administrators to choose
the way in which passwords are stored when a user's password is added or
set. The options are as follows:

legacy: Stores the password in R_USER_PASSWORD to match historical behavior.
Additionally, any passwords belonging to the user in the R_USER_CREDENTIALS
table will be cleared so older passwords can no longer be used. This is the
default behavior if the configuration is not present.

both: Stores the password in R_USER_PASSWORD to match historical behavior
and stores a salted hash of the user's password in R_USER_CREDENTIALS.

hashed: Stores a salted hash of the user's password in R_USER_CREDENTIALS.
Additionally, any passwords belonging to the user in the R_USER_PASSWORD
table will be cleared so that older passwords can no longer be used.

A new option has been added to the GeneralAdmin API which allows
the client to specify that a password for moduser is not scrambled. This
instructs the API to not attempt de-scrambling the password because it is
not in scrambled form. In order to maintain backwards-compatiblity, the
API changes the option given to the ModUser database operation to
"password-unobfuscated". This signals that the password should not be
de-scrambled before preparing it for insertion into the catalog.
Note: This option is required for servers which want to be FIPS compliant.