[ 8697] Add authentication session token support
This commit adds support for time-limited session tokens for use with
authentication. The intent is to allow authentication plugins to invoke
the new database operations for managing session tokens once a user has
authenticated with some authoritative service (e.g. checking a password
in the iRODS catalog). A session token is identified by a UUID which is
generated at creation time and can only be used by the user to which it
is issued.

The existing database table called R_USER_SESSION_KEY is being used to
store the session tokens as it is not known to be used by any server
features, plugins, or clients.

Session tokens are given expiration times which are represented as a
lifetime in seconds added to the epoch time at time of creation. This
lifetime is controlled by a grid configuration called
token_lifetime_in_seconds. If this configuration is not a valid value,
a default of 1209600 (2 weeks) is used. A rodsadmin may request a session
token which does not expire.

Session tokens belonging to a particular user may be deleted, and this
deletion can be narrowed down to include only expired tokens. This is
useful for developing processes whereby expired tokens can be regularly
cleaned up or a deleted user's tokens can be removed.

This commit also adds a new operation to the GeneralAdmin API to delete
session tokens called delete_session_tokens. This is accompanied by a
new database operation.