[ 8697] Add new "irods" authentication plugin
This adds a new iRODS authentication scheme called "irods". It is
intended to be a replacement for native authentication to which
deployments can migrate over time, or immediately. It is similar to
native authentication in that it uses user credentials stored in the
iRODS catalog.

This authentication plugin implements 3 server-side operations:
- server_prepare_auth_check: Performs some pre-flight checks to make
  sure that authentication with the connected client is allowed.
- server_auth_with_password: Takes the user's password and checks it
  with the stored hashes in R_USER_CREDENTIALS. If the password can be
  used to derive a matching key, a new session token is generated for
  the user, the user's expired session tokens are deleted, and the new
  session token is returned.
- server_auth_with_session_token: Takes a session token and checks it
  with the stored tokens in R_USER_SESSION_KEY. If the check is successful,
  the usual privilege flags are set in the RsComm and the connection
  is then authenticated.

The client-side operations allow for authenticating with either a password
or a session token. If password authentication is successful, the session
token returned by the server is recorded in a new file in the .irods
directory called .irods_secrets. The session token can then be used by
clients to authenticate instead of the user's password.