[ 8334] Fix potential leak in resolveRodsTarget
In resolveRodsTarget, when the destPath rodsPath_t is copied into targPath,
it is a shallow copy. This results in both structures pointing at the same
rodsObjStat_t. This causes problems for anybody using the function because
the targPath and destPath in the rodsPathInp_t must be freed in a complicated
manner since they are pointed at the same memory.

This commit addresses the issue by making a deep copy of the rodsPath_t. In
this way, the targPath owns its own rodsObjStat_t and can clean up in a
deterministic manner.